`urllib3`: move stubs for `ssl_match_hostname` around
