--- stubs/urllib3/urllib3/packages/ssl_match_hostname/__init__.pyi
@@ -1,4 +1,0 @@
-import ssl
-
-CertificateError = ssl.CertificateError
-match_hostname = ssl.match_hostname
