--- stubs/urllib3/urllib3/packages/ssl_match_hostname/_implementation.pyi
@@ -1,3 +1,0 @@
-class CertificateError(ValueError): ...
-
-def match_hostname(cert, hostname): ...
